Move space-related tests to new package
